requirements.txt
aniso8601==10.0.1
babel==2.17.0
bcrypt==4.3.0
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cssmin==0.2.0
cssselect2==0.8.0
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
Flask==2.3.3
Flask-Babel==2.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.58.4
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pillow==11.2.1
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
pyphen==0.17.2
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2022.7.1
redis==6.2.0
requests==2.32.4
rich==13.9.4
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.41
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.0
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.0.1
zopfli==0.2.3.post1
zxcvbn==4.5.0
